% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosme.R
\name{cosme}
\alias{cosme}
\title{Fit cosme model}
\usage{
cosme(model, data, option)
}
\arguments{
\item{model}{a \code{cosme_model} object}

\item{data}{a \code{data.frame}}

\item{option}{a \code{list} object with options passed to various methods. See details}
}
\description{
A short description...
}
\details{
Additional details...
}
