% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_cmv_cor.R
\name{estimate_cmv}
\alias{estimate_cmv}
\title{Estimate the Common Method Variance (CMV) coefficient of a set of variables}
\usage{
estimate_cmv(me_data)
}
\arguments{
\item{me_data}{a data frame or tibble of class \code{me}
which contains the desired variables from which to estimate the CMV.}
}
\value{
a numeric vector of length one with the estimated coefficient
}
\description{
Estimate the Common Method Variance (CMV) coefficient of a set of variables
}
\examples{
library(tibble)

me_df <-
 tibble(question = paste0("V", 1:5),
 quality = c(0.2, 0.3, 0.5, 0.6, 0.9),
 reliability = c(0.2, 0.4, 0.5, 0.5, 0.7),
 validity = c(0.5, 0.1, 0.6, 0.7, 0.8))

estimate_cmv(me_df)

\dontrun{

me_df <-
 tibble(question = paste0("V", 1:5),
 quality = c(0.2, 0.3, 0.5, 0.6, 0.9),
 reliability = c(0.2, 0.4, 0.5, 0.5, 0.7),
 validity = c(NA, 0.1, 0.6, 0.7, 0.8))

estimate_cmv(me_df)
# Error in estimate_cmv(me_df) :
# me_data must have non-missing values at columns reliability and validity for all variables
}


}
\seealso{
\code{\link{me_cmv_cor}} for automatically adjusting a correlation
matrix for the CMV, \code{\link{me_cmv_cov}} for automatically adjusting a
covariance matrix for the CMV.
}
