% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_rbind_metrics.R
\name{me_bind_metrics}
\alias{me_bind_metrics}
\title{Manually bind new rows to an me data frame}
\usage{
me_bind_metrics(me_data, question_name, metrics)
}
\arguments{
\item{me_data}{a data frame of class \code{me}}

\item{question_name}{a character string or a bare unquoted name that will be
used as the question name (not a column but a new question appended as row)}

\item{metrics}{a list containing new metrics. Currently it only
supports 'quality', 'reliability' and 'validity'. Can also specify one of the
metrics and the remaining are set to 0 by default}
}
\value{
\code{me_data} with a new row that contains the new question name
and the metrics as columns. Metrics must match existing names from SQP data
such as 'quality', 'reliability' and 'validity'.
}
\description{
Manually bind new rows to an me data frame
}
\examples{

# Toy dataset
library(tibble)

me_df <-
 tibble(question = paste0("V", 1:5),
 quality = c(0.2, 0.3, 0.5, 0.6, 0.9),
 reliability = c(NA, 0.4, 0.5, 0.5, 0.7),
 validity = c(NA, NA, 0.6, 0.7, 0.8))

me_df <- structure(me_df, class = c(class(me_df), "me"))

me_bind_metrics(me_df, new_question, list(quality = 0.7))
me_bind_metrics(me_df, new_question, list(quality = 0.7, reliability = 0.2))

# Specifying a wrong metric name results in error
\dontrun{
me_bind_metrics(me_df, new_question, list(wrong_metric = 0.7))
}

# Currently only quality, reliability and validity are allowed.

}
\seealso{
\code{\link{me_construct}} for the underlying workhorse.
}
