% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_correlate_covariance.R
\name{me_correlate}
\alias{me_correlate}
\title{Calculate a correlation matrix with an adjusted diagonal}
\usage{
me_correlate(
  .data,
  diag_adj = 1,
  wt = NULL,
  use = "complete.obs",
  method = "pearson"
)
}
\arguments{
\item{.data}{a matrix or data frame with numeric columns.}

\item{diag_adj}{a numeric vector with length equal to the number of columns
of \code{x} to be multiplied by the diagonal.  Alternatively, it can be of
length 1 which will be repeated through the whole diagonal.  If the
argument \code{wt} is used, then the length of \code{diag_adj} must be the
same as \code{x} excluding the weight column. By default it multiplies by
1, giving the same diagonal.}

\item{wt}{the name of the column which contains the weights as bare unquoted
names or as character vector or length 1. Note that when the weight
argument is specified, the estimation is done using
\code{\link[stats]{cov.wt}} instead of \code{\link[stats]{cor}} or
\code{\link[stats]{cov}}. This means that the arguments \code{use} and
\code{method} are ignored.}

\item{use}{an optional character string giving a method for computing
covariances in the presence of missing values. This must be (an
abbreviation of) one of the strings "everything", "all.obs",
"complete.obs", "na.or.complete", or "pairwise.complete.obs".}

\item{method}{a character string indicating which correlation coefficient (or
covariance) is to be computed. One of "pearson" (default), "kendall", or
"spearman": can be abbreviated.}
}
\value{
a correlation \code{tibble} with variable names as a column and the
  diagonal multiplied by \code{diag_adj}
}
\description{
\code{me_correlate} calculates a correlation matrix through
\code{\link[stats]{cor}} and multiplies the diagonal with the supplied
numeric vector. It's a wrapper around \code{\link[stats]{cor}} with slight
tweaks.
}
\examples{

# New diagonal
new_diagonal <- rnorm(ncol(mtcars))

me_correlate(mtcars)
me_correlate(mtcars, new_diagonal)
me_correlate(mtcars, new_diagonal, method = "kendall")

diagonal_wout_weight <- rnorm(ncol(mtcars) - 1)
me_correlate(mtcars, diagonal_wout_weight, wt = mpg)

}
\seealso{
\code{\link[stats]{cor}} for the workhorse behind the function and
  \code{\link[stats]{cov.wt}} for the function used for the weighting.
}
