% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_sscore.R
\name{me_sscore}
\alias{me_sscore}
\alias{me_sscore_}
\title{Calculate quality of sum score of selected variables}
\usage{
me_sscore(me_data, .data, new_name, ..., wt = NULL, .drop = TRUE)

me_sscore_(me_data, .data, new_name, vars_names, wt = NULL, .drop = TRUE)
}
\arguments{
\item{me_data}{a data frame of class \code{me} containing
quality estimates from the variables specified in \code{...}.}

\item{.data}{a data frame which contains data for the variables specified
in \code{...}.}

\item{new_name}{a bare unquoted name or a string specifying the name
of the new sum score.}

\item{...}{bare unquoted names or separate strings specifying
the variable names from which to estimate quality of their sum score.
They all must be present in \code{me_data} and \code{df}. At minimum,
it must be two or more variable names.}

\item{wt}{a non-NA numeric vector of the same length as the variables
specified in \code{...}. This will be used as weights in calculating the
sum scores of all variable. Be default, all variables are given the same
weight.}

\item{.drop}{a logical stating whether to drop the questions that compose
the sum score (specified in \code{...}) If \code{FALSE} it retains the
original questions and the composite score.}

\item{vars_names}{character vector specifying the variable names from
which to estimate quality of their sum score. They all must be present
in \code{me_data} and \code{df}. At minimum, it must be two or more
variable names.}
}
\value{
a \code{\link[tibble]{tibble}} similar to \code{me_data} but
with a new row containing the quality of a sum score with the name
specified in \code{new_name}.
}
\description{
\code{me_sscore} takes a data frame with quality estimates of class
\code{me} and estimates the quality of a sum score for
selected variables in \code{...}. Use \code{me_sscore_} if you're
interested in programming with \code{me_sscore}
}
\examples{
# Political trust example
data(ess7es)

quality <-
  data.frame(
    question = c("trstprl", "trstplt", "trstprt"),
    reliability = c(0.901110426085505, 0.923038460737146, 0.926282894152753),
    validity = c(0.979285453787607, 0.982344135219425, 0.977752524926425),
    quality = c(0.882609766544649, 0.906642156531451, 0.906090503205943)
  )

selected_vars <- c("trstprl", "trstplt", "trstprt")
score <- me_sscore(quality[quality$question \%in\% selected_vars, ],
                   ess7es,
                   new_name = "s1",
                   trstprl, trstplt, trstprt,
                   wt = NULL)

# Returns the quality and method effect of any given sum score
score

# State services example
quality <-
  data.frame(
    question = c("stfedu", "stfhlth"),
    reliability = c(0.870057469366248, 0.871779788708135),
    validity = c(0.915423399307664, 0.893308457365092),
    quality = c(0.796868872525461, 0.779102047231298)
  )

score <- me_sscore(quality,
                   ess7es,
                   new_name = "s1",
                   stfedu, stfhlth,
                   wt = NULL)

# Returns the quality and method effect of any given sum score
score


}
