% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosme.R
\name{print.cosme_est}
\alias{print.cosme_est}
\title{Print Estimates from \code{comse_est}}
\usage{
\method{print}{cosme_est}(x, ..., nd = 3L)
}
\arguments{
\item{...}{Further arguments passed to or from other methods}

\item{nd}{Number of decimal places to print (Defaults 3)}

\item{object}{a \code{cosme_model} object}
}
\description{
A short description...
}
