% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosme.R
\name{summary.cosme_fit}
\alias{summary.cosme_fit}
\title{Summarize from \code{comse_fit}}
\usage{
\method{summary}{cosme_fit}(object, ...)
}
\arguments{
\item{object}{a \code{cosme_model} object}

\item{...}{additional arguments passed to \code{\link{print}}}
}
\description{
A short description...
}
