% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosme.R
\name{cosme}
\alias{cosme}
\title{Fit cosme model}
\usage{
cosme(model, data, ...)
}
\arguments{
\item{model}{a \code{cosme_model} object}

\item{data}{a \code{data.frame}}

\item{...}{passed onto other stuff}
}
\description{
A short description...
}
