% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosme.R
\name{cosme}
\alias{cosme}
\title{Fit cosme model}
\usage{
cosme(model, data, info = FALSE, option)
}
\arguments{
\item{model}{a \code{cosme_model} object}

\item{data}{a \code{\link{data.frame}}}

\item{info}{a \link{logical} If \code{TRUE} then fit information theory}

\item{option}{a \code{\link{list}} object with options passed to various methods. See details}
}
\description{
A short description...
}
\details{
Additional details...
}
