% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_cmv_cor.R
\name{estimate_cmv}
\alias{estimate_cmv}
\title{Estimate the Common Method Variance (CMV) coefficient of a set of variables}
\usage{
estimate_cmv(me_data)
}
\arguments{
\item{me_data}{a data frame or tibble of class \code{me}
which contains the desired variables from which to estimate the CMV.}
}
\value{
a numeric vector of length one with the estimated coefficient. The
result is a named vector where the name indicates the CMV between the
two variables.
}
\description{
Estimate the Common Method Variance (CMV) coefficient of a set of variables
}
\examples{
library(tibble)

# Don't remove don't run: since this fun is not exported
# it throws an error. Keeping docs jsut for me here.
\dontrun{

# Example with political trust
me_df <-
  data.frame(
    question = c("trstprl", "trstplt", "trstprt"),
    reliability = c(0.901110426085505, 0.923038460737146, 0.926282894152753),
    validity = c(0.979285453787607, 0.982344135219425, 0.977752524926425),
    quality = c(0.882609766544649, 0.906642156531451, 0.906090503205943),
    method_eff = c(0.1824610, 0.1726847, 0.1942987)
  )

# CMV between each variable
estimate_cmv(me_df)

# State services example

me_df <-
  data.frame(
    question = c("stfedu", "stfhlth"),
    reliability = c(0.870057469366248, 0.871779788708135),
    validity = c(0.915423399307664, 0.893308457365092),
    quality = c(0.796868872525461, 0.779102047231298),
    method_eff = c(0.3501914, 0.3918163)
  )

# CMV between these two variables
estimate_cmv(me_df)

}

}
\seealso{
\code{\link{me_cmv_cor}} for automatically adjusting a correlation
matrix for the CMV, \code{\link{me_cmv_cov}} for automatically adjusting a
covariance matrix for the CMV.
}
