% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_cmv_cor.R
\name{me_cmv_cor}
\alias{me_cmv_cor}
\title{Adjust a correlation matrix for Common Method Variance (CMV)}
\usage{
me_cmv_cor(.medesign, cmv = NULL)
}
\arguments{
\item{.medesign}{An \code{medesign} object given by \code{\link{medesign}}}

\item{cmv}{an optional numeric vector of the same length as the number of
common method variance specifications in \code{.mdesign}.This argument is
left available if the user has reasons to input their own CMV. By default,
it is set to NULL and it is calculated internally.}
}
\value{
The common-method-variance corrected correlation.
}
\description{
\code{me_cmv_cor} accepts an \code{medesign} object specified in
\code{\link{medesign}} and adjusts the correlation coefficients of
common method variables with the reliability and validity coefficients
from \code{me_data}. Optionally, you can supply the CMV coefficients
manually in the argument \code{cmv}.
}
\examples{

set.seed(2131)
library(tibble)

original_df <- as.data.frame(matrix(rnorm(100, sd = 50), nrow = 20))

# Toy quality dataset
me_df <-
 tibble(question = paste0("V", 1:5),
 reliability = c(0.6, 0.4, 0.5, 0.5, 0.7),
 validity = c(0.79, 0.9, 0.6, 0.7, 0.8),
 quality = c(0.2, 0.3, 0.5, 0.6, 0.9))

# Define mdesign object
m_obj <- medesign("~ V4 + V5", original_df, me_df)

# Original correlation matrix
me_correlate(original_df)

# Coefficient of correlation changes
# when adjusting for common method variance
me_cmv_cor(m_obj)

# The V5*V4 from both the upper/lower triangles
# correlation matrix changed from -0.05 to -0.27

}
\seealso{
\code{link{medesign}} and \code{\link{me_cmv_cov}} for the same
adjustment but for a covariance matrix.
}
