% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_cmv_cor.R
\name{me_cmv_cor}
\alias{me_cmv_cor}
\title{Adjust a correlation matrix for Common Method Variance (CMV) and Quality of Sum score}
\usage{
me_cmv_cor(.medesign)
}
\arguments{
\item{.medesign}{An \code{medesign} object given by \code{\link{medesign}}}
}
\value{
The common-method-variance corrected correlation. It is a correlation
data frame but some correlation coefficients are adjusted for their shared
common method variance and the quality of sum scores.
}
\description{
\code{me_cmv_cor} accepts an \code{medesign} object created by
\code{\link{medesign}} and adjusts the correlation coefficients of
common method variables with the reliability and validity coefficients
from \code{me_data}.
}
\examples{

data(ess7es)

.data <- ess7es[c("trstplt", "trstprl", "trstprt")]

# Correct for measurement error for all variables
# and correct for the fact that these three variables
# share a common method
me_syntax <- "~~ .; ~ trstplt + trstprl + trstprt"

# Data from sqp
me_data <-
  data.frame(
    question = c("trstprl", "trstplt", "trstprt"),
    reliability = c(0.812, 0.852, 0.858),
    validity = c(0.959, 0.965, 0.956),
    quality = c(0.779, 0.822, 0.821)
  )

mdes <- medesign(me_syntax, .data, me_data)

# Correlations adjusted for CMV
me_cmv_cor(mdes)

# Original correlation
me_correlate(.data)

# More elaborate with two pairs of variables
# sharing common methods

# Each pair of variables share a common method.
me_syntax <-
 "~~ .
  ~ trstplt + trstprl + trstprt
  ~ stfedu + stfhlth"

me_data2 <-
  data.frame(
    question = c("stfedu", "stfhlth"),
    reliability = c(0.870057469366248, 0.871779788708135),
    validity = c(0.915423399307664, 0.893308457365092),
    quality = c(0.796868872525461, 0.779102047231298)
  )

me_data <- rbind(me_data, me_data2)

.data <- ess7es[c("trstplt", "trstprl", "trstprt", "stfedu", "stfhlth")]
mdes <- medesign(me_syntax, .data, me_data)

# Correlations adjusted for CMV
me_cmv_cor(mdes)

# Original correlations
me_correlate(.data)

}
\seealso{
\code{link{medesign}} and \code{\link{me_cmv_cov}} for the same
adjustment but for a covariance matrix.
}
