% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_cmv_cor.R
\name{me_cmv_cor}
\alias{me_cmv_cor}
\title{Adjust a correlation matrix for Common Method Variance (CMV)}
\usage{
me_cmv_cor(x, me_data, ..., cmv = NULL)
}
\arguments{
\item{x}{a correlation matrix, a correlation data frame or a correlation
\code{tibble} given by \code{\link{me_correlate}}.}

\item{me_data}{a data frame or tibble of class \code{me}.}

\item{...}{two or more variables present in both \code{x} and \code{me_data}. Can
be both in bare unquoted names or as character strings.}

\item{cmv}{an optional numeric vector of length 1 which contains the
CMV coefficient of the variables specified in \code{...}.
This argument is left available if the user has reasons to input their own CMV.
By default, it is set to NULL and it is calculated internally.}
}
\value{
the same matrix supplied in \code{x} but as a tibble with
the correlation coefficients of the variables supplied in \code{...}
adjusted for their common method variance.
If \code{x}  had row names, they're moved as a column called
'rowname'. If it doesnt have row names or a column named 'rowname' a new
column is created with the column names to mimic a correlation matrix.
}
\description{
\code{me_cmv_cov} accepts a correlation matrix, a correlation data frame or a correlation
tibble from \code{\link{me_correlate}}
and adjusts the coefficients of the variables specified
in  \code{...} with the reliability and validity coefficients from \code{me}.
All variables specified in \code{...} must be present in both \code{x}
and \code{me_data}. Optionally, you can supply the cmv coefficient in the
argument \code{cmv}.
}
\examples{

set.seed(2131)
library(tibble)

original_df <- as.data.frame(matrix(rnorm(100, sd = 50), nrow = 20))
corr_tibble <- me_correlate(original_df, rnorm(5))

# Toy dataset
me_df <-
 tibble(question = paste0("V", 1:5),
 quality = c(0.2, 0.3, 0.5, 0.6, 0.9),
 reliability = c(NA, 0.4, 0.5, 0.5, 0.7),
 validity = c(NA, NA, 0.6, 0.7, 0.8))

me_df <- structure(me_df, class = c(class(me_df), "me"))

# Show that when y is not from me class, there's an error

# Original correlation matrix with new diagonal
corr_tibble

# Coefficient of correlation changes
# when adjusting for common method variance
me_cmv_cor(corr_tibble, me_df, V4, V5)

# The V5*V4 from both the upper/lower triangles
# correlation matrix changed from -0.05 to -0.203

}
\seealso{
\code{\link{me_correlate}} for calculating correlation matrices
and \code{\link{estimate_cmv}} for calculating the CMV manually.
}
