% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_parse_model.R
\name{me_parse_model}
\alias{me_parse_model}
\title{Extract a model specification from the measurement error model}
\usage{
me_parse_model(
  model_syntax = " ",
  me_data = NULL,
  data = NULL,
  as_list = FALSE,
  debug_code = FALSE,
  operators = c("~", "=", "~~")
)
}
\arguments{
\item{model_syntax}{a string containing the model specification}

\item{me_data}{measurment error data. This is used to return the RHS of
the quality specification.}

\item{data}{data to extract the variables when specifying '~ .'}

\item{as_list}{a logical for whether the result is a data frame or a list.
Dataframe is set as a default.}

\item{debug_code}{a logical for whether debugging logs should be printed.}

\item{operators}{permitted operators to define new lines}
}
\value{
A data frame with the parsed model extracted into different columns
}
\description{
Extract a model specification from the measurement error model
}
